move empty_keyring fixture to conftest
